app/requirements.txt
discord.py
python-dotenv
trueskill